made hash size smaller

+++ src/rust/src/shingleset.rs
@@ -6,14 +6,14 @@ use fxhash::FxHasher;
 
 #[derive(Debug, Clone)]
 pub struct ShingleSet {
-    pub shingles: HashSet<u32>,
+    pub shingles: HashSet<u16>,
     pub shingle_len : usize,
     pub index : usize,
 }
 
 impl ShingleSet {
     pub fn new(string: &String, shingle_len: usize, index: usize) -> Self {
-        let mut out_set: HashSet<u32> = HashSet::new();
+        let mut out_set: HashSet<u16> = HashSet::new();
 
         let char_vec: Vec<char> = string.chars().collect();
 
@@ -22,7 +22,7 @@ impl ShingleSet {
 
             window.hash(&mut hasher);
 
-            let result: u32 = hasher.finish() as u32;
+            let result: u16 = hasher.finish() as u16;
 
             out_set.insert(result);
         }
